requirements.txt
Django==1.9.7
django-cors-headers==1.1.0
django-debug-toolbar==1.4
django-extensions==1.6.7
django-filter==0.13.0
django-material==0.12.3
djangorestframework==3.3.3
djangorestframework-filters==0.8.0
django-autoslug==1.9.3
et-xmlfile==1.0.1
jdcal==1.2
Markdown==2.6.2
Pillow==3.2.0
six==1.10.0
sqlparse==0.1.19
